+++ src/main/kotlin/xyz/acevedosharp/ui_models/Lote.kt
@@ -1,6 +1,5 @@
 package xyz.acevedosharp.ui_models
 
-import javafx.beans.property.SimpleDoubleProperty
 import javafx.beans.property.SimpleIntegerProperty
 import javafx.beans.property.SimpleObjectProperty
 import tornadofx.ItemViewModel
@@ -8,12 +7,12 @@ import tornadofx.getValue
 import tornadofx.setValue
 import xyz.acevedosharp.persistence.entities.ProductoDB
 
-class Lote(var id: Int?, cantidad: Int, precioCompra: Double, producto: ProductoDB) {
+class Lote(var id: Int?, cantidad: Int, precioCompra: Int, producto: ProductoDB) {
     val cantidadProperty = SimpleIntegerProperty(this, "cantidad", cantidad)
     var cantidad: Int by cantidadProperty
 
-    val precioCompraProperty = SimpleDoubleProperty(this, "precioCompra", precioCompra)
-    var precioCompra: Double by precioCompraProperty
+    val precioCompraProperty = SimpleIntegerProperty(this, "precioCompra", precioCompra)
+    var precioCompra: Int by precioCompraProperty
 
     val productoProperty = SimpleObjectProperty(this, "producto", producto)
     var producto: ProductoDB by productoProperty
